Created the Block and Chunk Class
In the chunk class, inside of the constructor that accepts the x, y, and
z values (At the bottom of class, pretty much the only error in it),
there is a  section that tries to create BlockType_Default. Now, I was
thinking I would just change that to Grass, but I left it there so you
would see it. Let me know!