Added in more blocks
I added the 6 different blocks into the chunk constructor with each
having a 1/6 chance of being picked. I decided default should just be
bedrock. I tried adding the Chunk.render method to the Camera class but
ended up getting an error. I sent an email to the professor about it, so
if I get a response I'll change it. Otherwise you can take a look.