Moved Frustrum methods to the chunk class. Unbound the light from camera movement.
I can't figure out where to put the Frustrum culling method, since the
render method for chunk is only called once.
Would drawing be a different step from rendering and placing them in the
scene? I think I may have to code in such a way that the chunk is being
rendered again each frame, and only those in the frustrum, but that
doesn't sound like an optimization to me.